hibernate完成版

+++ UserManage/src/main/java/com/wq/data/UserRepository.java
@@ -5,9 +5,9 @@ import com.wq.*;
 public interface UserRepository {
 	List<User> getAll();
 	User findByName(String name);
-	User findById(Long id);
+	User findById(Integer id);
 	void save(User u);
 	boolean LoginRight(User u);
 	boolean Registed(String name);
-	void deleteById(long id);
+	void deleteById(Integer id);
 }
